+++ src/project/__init__.py
@@ -1,0 +1,6 @@
+"""项目管理模块"""
+
+from .manager import ProjectManager
+from .models import Project, ProjectError, ProjectErrorCode
+
+__all__ = ["ProjectManager", "Project", "ProjectError", "ProjectErrorCode"]